refactor(core): add IsFatal filter, improve error handling
Introduce ExceptionFilters.IsFatal for fatal exception detection and update all catch blocks to use it, ensuring fatal exceptions propagate. Refactor disposal and batching logic for clarity and reliability. Improve error handling strategies, event handler registration, and test code correctness.

+++ HoneyDrunk.Transport/HoneyDrunk.Transport.AzureServiceBus/Mapping/EnvelopeMapper.cs
@@ -180,7 +180,7 @@ public static class EnvelopeMapper
     /// <returns>The delivery count.</returns>
     public static int GetDeliveryCount(ServiceBusReceivedMessage message)
     {
-        return (int)message.DeliveryCount;
+        return message.DeliveryCount;
     }
 
     /// <summary>
